[life-xp-xp-service-30] Finish addXpRecord with E2E tests -second part

+++ life-xp-xp-service-core/src/main/java/com/bladestepapp/lifexpxpservicecore/domain/XpActivity.java
@@ -13,9 +13,9 @@ public class XpActivity {
 
     private UUID activityId;
 
-    private Integer xp;
+    private long xp;
 
-    public static XpActivity create(UUID userId, UUID activityId, Integer xp){
+    public static XpActivity create(UUID userId, UUID activityId, long xp){
         return XpActivity.builder()
                 .userId(userId)
                 .activityId(activityId)
